Adds Baidu support
This commit also changes the regular expression used to capture the
provider domain. I have mixed feelings about this and it will likely
change again in the future. For some reason, the previous version
properly captured the bing domain but not the baidu domain. I spent
time looking at it but it wasn't obvious. I'm sure it will be at some point.